Sniper public/requirements.txt
tensorflow>=2.10.0
keras>=2.10.0
pandas>=1.5.0
numpy>=1.23.0
tqdm>=4.64.0
docopt>=0.6.2
coloredlogs>=15.0
python-dotenv>=0.21.0
solana>=0.27.0
python-binance>=1.0.16
web3>=5.31.1